# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Servanda Office API
  version: 1.0.0
  description: |
    Multi-Tenant LegalTech Platform API for contract creation and management.

    **Core Flow:** Template selection -> Guided Q&A Interview -> Conflict-free Contract -> DOCX/ODT Export.

    **Architecture Decisions:**
    - ADR-001: Shared DB with PostgreSQL Row-Level Security for tenant isolation
    - ADR-002: Immutable versioning for Clauses/Templates; ContractInstance pins versions
    - ADR-003: Async export via pgboss job queue with S3 storage
    - ADR-004: Server-side ODT conversion via LibreOffice headless

    **Authentication:**
    - Production: Bearer JWT (OIDC/SAML)
    - Development: Header-based tenant context (x-tenant-id, x-user-id, x-user-role)

    **Pagination:** All list endpoints support `page` and `pageSize` query parameters.
    Default page size: 20, maximum: 100.

    **Status Workflow (Clause/Template Versions):**
    `draft` -> `review` -> `approved` -> `published` -> `deprecated`
    (review -> draft and approved -> draft are also valid transitions)
  contact:
    name: Team 01 - Product Architecture
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server (v1 canonical prefix)
  - url: http://localhost:3000/api
    description: Legacy route prefix (backward-compatible)
security:
  - BearerAuth: []
  - DevHeaders: []
...
paths:
...
  /export-jobs/{id}/download:
    get:
      tags: [Export]
      summary: Download exported document
      description: |
        Redirects to the pre-signed download URL for a completed export.
        Only available when the export job status is `done`.
      operationId: downloadExport
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '302':
          description: Redirect to download URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Export not yet complete or has failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # =========================================================================
  # DLQ (Dead-Letter Queue)
  # =========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Production authentication via OIDC/SAML JWT token.
        Include tenant claim in the JWT payload.
    DevHeaders:
      type: apiKey
      in: header
      name: x-tenant-id
      description: |
        Development-mode authentication via headers.
        Requires all three headers:
        - `x-tenant-id` (UUID) - Tenant identifier
        - `x-user-id` (UUID) - User identifier
        - `x-user-role` (admin|editor|user) - User role

  # =========================================================================
  # Parameters
  # =========================================================================
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # =========================================================================
  # Schemas
  # =========================================================================
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          example: Resource not found
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...